requirements.txt
fastapi~=0.112.0
requests
uvicorn
pydantic~=2.8.2
python-dotenv~=1.0.1
gunicorn